# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/twin/{twin_id}": {
            "post": {
                "description": "submit the message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "submit the message",
                "parameters": [
                    {
                        "description": "rmb.Message",
                        "name": "msg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rmbproxy.Message"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "twin id",
                        "name": "twin_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rmbproxy.MessageIdentifier"
                        }
                    }
                }
            }
        },
        "/twin/{twin_id}/{retqueue}": {
            "get": {
                "description": "Get the message result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "Get the message result",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "twin id",
                        "name": "twin_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "message retqueue",
                        "name": "retqueue",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rmbproxy.Message": {
            "type": "object",
            "properties": {
                "cmd": {
                    "type": "string",
                    "example": "zos.statistics.get"
                },
                "dat": {
                    "type": "string"
                },
                "dst": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "err": {
                    "type": "string"
                },
                "exp": {
                    "type": "integer",
                    "example": 0
                },
                "now": {
                    "type": "integer",
                    "example": 1631078674
                },
                "ret": {
                    "type": "string"
                },
                "shm": {
                    "type": "string"
                },
                "src": {
                    "type": "integer",
                    "example": 1
                },
                "try": {
                    "type": "integer",
                    "example": 2
                },
                "uid": {
                    "type": "string"
                },
                "ver": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "rmbproxy.MessageIdentifier": {
            "type": "object",
            "properties": {
                "retqueue": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                }
            }
        }
    }
}